+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/_internal/state/HubdleState.kt
@@ -315,6 +315,7 @@ internal data class HubdleState(
         data class LanguageSettings(
             var experimentalContracts: Boolean = false,
             var experimentalCoroutinesApi: Boolean = false,
+            var experimentalSerializationApi: Boolean = true,
             var experimentalStdlibApi: Boolean = true,
             var experimentalTime: Boolean = true,
             var flowPreview: Boolean = false,